Added: checkout flow, but it's bad

+++ internal/dto/product.go
@@ -5,7 +5,7 @@ import "time"
 type CreateProductRequest struct {
 	Name        string
 	Description string
-	Price       float64
+	Price       int64
 	Slug        string
 }
 
@@ -13,14 +13,14 @@ type ProductResponse struct {
 	ID          int
 	Name        string
 	Description string
-	Price       float64
+	Price       int64
 	Slug        string
 	CreatedAt   time.Time
 }
 
 type UpdateProductRequest struct {
-	Name        *string  `json:"name"`
-	Description *string  `json:"description"`
-	Price       *float64 `json:"price"`
-	Slug        *string  `json:"slug"`
+	Name        *string `json:"name"`
+	Description *string `json:"description"`
+	Price       *int64  `json:"price"`
+	Slug        *string `json:"slug"`
 }
